requirements.txt
runpod==1.3.0
torch==2.0.1
transformers>=4.30.0
diffusers>=0.19.0
accelerate>=0.20.3
Pillow>=9.5.0
opencv-python>=4.7.0
numpy>=1.24.3
xformers==0.0.20
safetensors>=0.3.1
huggingface-hub>=0.15.1
